Original change: https://googleplex-android-review.googlesource.com/c/platform/frameworks/native/+/15013550

+++ services/surfaceflinger/Layer.cpp
@@ -2020,7 +2020,7 @@ int32_t Layer::getBackgroundBlurRadius() const {
 
 const std::vector<BlurRegion> Layer::getBlurRegions() const {
     auto regionsCopy(getDrawingState().blurRegions);
-    int layerAlpha = getAlpha();
+    float layerAlpha = getAlpha();
     for (auto& region : regionsCopy) {
         region.alpha = region.alpha * layerAlpha;
     }
